Move BionicDrivetrain.java
