+++ src/org/team4909/bionic/utils/commands/DriveOI.java
@@ -1,8 +1,8 @@
 package org.team4909.bionic.utils.commands;
 
-import org.team4909.bionic.utils.subsystems.BionicDrivetrain;
-
 import edu.wpi.first.wpilibj.command.Command;
+
+import org.team4909.bionic.utils.drivetrain.BionicDrivetrain;
 
 public class DriveOI extends Command {
 	private BionicDrivetrain drivetrainSubsystem;
