+++ src/org/team4909/preseason2018/core/OI.java
@@ -2,9 +2,9 @@ package org.team4909.preseason2018.core;
 
 import org.team4909.bionic.utils.commands.SetDoubleSolenoid;
 import org.team4909.bionic.utils.commands.SetDriveDirection;
+import org.team4909.bionic.utils.drivetrain.BionicDrivetrain.Direction;
 import org.team4909.bionic.utils.oi.BionicAxis;
 import org.team4909.bionic.utils.oi.BionicF310;
-import org.team4909.bionic.utils.subsystems.BionicDrivetrain.Direction;
 
 import edu.wpi.first.wpilibj.DoubleSolenoid.Value;
